mara-hcp/simulators/market-sim/requirements.txt
kafka-python==2.0.2
numpy==1.26.3
asyncio==3.4.3

